refactor(gfql/polars): home the polars backend under lazy/engine/polars/ (was engine_polars/)
Pure move, no logic change. lazy/__init__ already specifies per-backend lowering under
lazy/engine/<backend>/ (future lazy/engine/duckdb, .../dask); the eager polars lowering
was in a flat sibling engine_polars/. Collapse it into the one polars-backend home so
future lazy backends don't proliferate flat engine_* siblings:

- engine_polars/{chain,dtypes,predicates,projection,row_pipeline}.py -> lazy/engine/polars/
- engine_polars/hop.py (eager) -> lazy/engine/polars/hop_eager.py (distinct name; the existing
  lazy collect-once hop.py is left in place — no collision).
- engine_polars/__init__ re-exports folded into lazy/engine/polars/__init__; engine_polars/ removed.
- imports rewritten engine_polars -> lazy.engine.polars (eager hop -> .hop_eager).
