+++ graphistry/compute/chain.py
@@ -805,7 +805,7 @@ def chain(
         # plans/gfql-polars-engine). Correctness gated by differential parity.
         if validate_schema:
             Chain(ops if not isinstance(ops, Chain) else ops.chain).validate(collect_all=False)
-        from graphistry.compute.gfql.engine_polars.chain import chain_polars
+        from graphistry.compute.gfql.lazy.engine.polars.chain import chain_polars
         # NO pandas fallback here (see plan.md NO-CHEATING): chain_polars raises
         # NotImplementedError for deferred features (var-length/multi-hop edges,
         # undirected multi-edge); that honest signal propagates to the caller.
