+++ graphistry/compute/gfql/lazy/engine/polars/predicates.py
